MPT-16303 Fixes for broken tests (#303)
* Update resources page mock filter data

Expanded and updated the mock data for AvailableFiltersResponse in resources-page.mocks.ts to include more filter values and details. This change supports more comprehensive e2e testing scenarios for the resources page.

* Increase timeout for resource item wait in tests

Set a 15-second timeout for waiting on the first resource item in the resources page tests to improve test reliability.